vai_paqueta_backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.2
django-cors-headers==4.4.0
geopy==2.4.1
psycopg2-binary==2.9.9
